Move OpenAPIWebhookExample
